// src/app/home/interfaces/movieCast.interface.ts
export interface MovieCastResAPI {
    imageBaseUrl: string;
    data:         Datum[];
}

export interface Datum {
    adult:                boolean;
    gender:               number;
    id:                   number;
    known_for_department: KnownForDepartment;
    name:                 string;
    original_name:        string;
    popularity:           number;
    profile_path:         null | string;
    cast_id:              number;
    character:            string;
    credit_id:            string;
    order:                number;
}

export enum KnownForDepartment {
    Acting = "Acting",
    Crew = "Crew",
}